Simplify text scanner
Use a regex instead of `index` method.

The article instead delegates to the SimpleScanner (which parses the symbols)
in order to implement the TextScanner. I decided not to implement it this way
because it feels very unintutive to me, and I did not implement the SimpleScanner
in the same way (as a single object with a hash lookup)