Extract token matching to `PatternBuilder`
Previously the parser had to know about Tokens, token
predicate objects, and parse tree nodes, which was quite confusing.
I think delegating to another object makes the parser a bit simpler
to understand.

I've moved this to the tokenizer file as it seems more related to the
tokenisation than the parsing.