Add generator to print text to standard out
I decided to implement this instead of the HTML generator example from
the blog post[1].

I thought I would find it more useful to have a command line tool which
could show rich text in a terminal.

I also found the example of the visitor pattern[2] in the blog post to
be quite weird. There, the logic of traversing the tree is
coupled with the HTML generation, and there is a seperate class for each
part of the AST.

I've instead seperated these two concerns into 2 classes so that the
`Generator` is responsible for calling methods on the `ConsoleOutput`
visitor.

[1]. https://blog.beezwax.net/2018/05/25/writing-a-markdown-compiler
[2]. https://en.wikipedia.org/wiki/Visitor_pattern